+++ uMockingSuite.Tests/obj/Debug/net10.0/uMockingSuite.Tests.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("uMockingSuite.Tests")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+96c2d4cf0c9f7ff10185550167d6d13564be5a85")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+6a35bb3e13c46b9e64f23bfd539e90f673d40462")]
 [assembly: System.Reflection.AssemblyProductAttribute("uMockingSuite.Tests")]
 [assembly: System.Reflection.AssemblyTitleAttribute("uMockingSuite.Tests")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]