+++ uMockingSuite/Services/IMockingService.cs
@@ -5,4 +5,5 @@ namespace uMockingSuite.Services;
 public interface IMockingService
 {
     string GetMockingMessage(IContent content);
+    Task<string> GetMockingMessageAsync(string contentName, string contentTypeAlias);
 }